// src/app/feature/model/employees.model.ts
export interface Employees {
    id: number;
    userId: number;
    name: string;
    phone:string;
    email: string;
    dateOfBirth: string;
    genderId: number;
    address: string;
    basicSalary: string;
    mobileNumber: string;
    status:boolean;
    joiningDate:string;
    resignDate:string;
    departmentId: number;
    designationId: number;
    acountName: string;
    acountNumber: string;
    swiftCode: string;
    brance: string;
   
  }